requirements.txt
PyYAML>=6.0
python-dateutil>=2.9
jinja2>=3.0
markdown>=3.3